Fix more -Wconversion warnings
We still have many -Wconversion warnings left. Though about half of them
originate from only 4 files. Mostly because it's in templated code, and
that template gets instantiated many times.

This patch fixes all warnings in:
    YM2413NukeYKT
    SaI3xScaler
    VDPCmdEngine
    V9990CmdEngine

+++ src/sound/YM2413NukeYKT.hh
@@ -184,7 +184,7 @@ private:
 	template<uint32_t CYCLES>                               ALWAYS_INLINE void doRegWrite();
 	template<uint32_t CYCLES>                               ALWAYS_INLINE void doIO();
 
-	NEVER_INLINE void doRegWrite(uint32_t channel);
+	NEVER_INLINE void doRegWrite(uint8_t channel);
 	             void doRegWrite(uint8_t block, uint8_t channel, uint8_t data);
 	NEVER_INLINE void doIO(uint32_t cycles_plus_1, Write& write);
